Windows, Linux consistent app/window menu, centralized functions with MacOS
