+++ src/main/createWindow.ts
@@ -6,18 +6,19 @@
 // ==LICENSE-END=
 
 import * as debug_ from "debug";
-import { app, BrowserWindow, Menu, shell, webContents } from "electron";
+import { app, BrowserWindow, Menu, shell } from "electron";
 import * as path from "path";
 import { AppWindowType } from "readium-desktop/common/models/win";
 import {
     getWindowsRectangle,
 } from "readium-desktop/common/rectangle/window";
-import { Translator } from "readium-desktop/common/services/translator";
 import { container } from "readium-desktop/main/di";
 import { WinRegistry } from "readium-desktop/main/services/win-registry";
 import {
     _PACKAGING, _RENDERER_APP_BASE_URL, IS_DEV,
 } from "readium-desktop/preprocessor-directives";
+
+import { setMenu } from "./menu";
 
 // Logger
 const debug = debug_("readium-desktop:createWindow");
@@ -73,10 +74,7 @@ export async function createWindow() {
 
     mainWindow.loadURL(rendererBaseUrl);
 
-    // Create the app menu on mac os to allow copy paste
-    if (process.platform === "darwin") {
-        initDarwin();
-    }
+    setMenu(mainWindow);
 
     if (IS_DEV) {
         const {
@@ -93,9 +91,6 @@ export async function createWindow() {
 
         // Open dev tools in development environment
         mainWindow.webContents.openDevTools();
-    } else {
-        // Remove menu bar
-        mainWindow.setMenu(null);
     }
 
     // Redirect link to an external browser
@@ -129,176 +124,3 @@ app.on("activate", async () => {
         await createWindow();
     }
 });
-
-export function initDarwin() {
-    const translator = container.get("translator") as Translator;
-    const template: Electron.MenuItemConstructorOptions[] = [
-        {
-            label: "Thorium",
-            submenu: [
-                {
-                    role: "togglefullscreen",
-                },
-                {
-                    role: "minimize",
-                },
-                {
-                    role: "close",
-                },
-                {
-                    role: "front",
-                },
-                {
-                    type: "separator",
-                },
-                {
-                    role: "hide",
-                },
-                {
-                    role: "hideothers",
-                },
-                {
-                    role: "unhide",
-                },
-                {
-                    type: "separator",
-                },
-                {
-                    role: "services",
-                    submenu: [],
-                },
-                {
-                    type: "separator",
-                },
-                {
-                    role: "quit",
-                    // accelerator: "Command+Q",
-                    // click: () => { app.quit(); },
-                    label: translator.translate("app.quit"),
-                },
-            ],
-        },
-        {
-            label: translator.translate("app.edit.title"),
-            role: "edit",
-            submenu: [
-                {
-                    role: "undo",
-                    // accelerator: "CmdOrCtrl+Z",
-                    // selector: "undo:",
-                    label: translator.translate("app.edit.undo"),
-                },
-                {
-                    role: "redo",
-                    // accelerator: "Shift+CmdOrCtrl+Z",
-                    // selector: "redo:",
-                    label: translator.translate("app.edit.redo"),
-                },
-                {
-                    type: "separator",
-                },
-                {
-                    role: "cut",
-                    // accelerator: "CmdOrCtrl+X",
-                    // selector: "cut:",
-                    label: translator.translate("app.edit.cut"),
-                },
-                {
-                    role: "copy",
-                    // accelerator: "CmdOrCtrl+C",
-                    // selector: "copy:",
-                    label: translator.translate("app.edit.copy"),
-                },
-                {
-                    role: "paste",
-                    // accelerator: "CmdOrCtrl+V",
-                    // selector: "paste:",
-                    label: translator.translate("app.edit.paste"),
-                },
-                {
-                    role: "selectall",
-                    // accelerator: "CmdOrCtrl+A",
-                    // selector: "selectAll:",
-                    label: translator.translate("app.edit.selectAll"),
-                },
-            ],
-        },
-    ];
-    if (IS_DEV) {
-        template.push(
-            {
-                label: "DEV",
-                submenu: [
-                    {
-                        label: "RELOAD WINDOW",
-                        accelerator: "CmdOrCtrl+R",
-                        click: (_item: any, focusedWindow: any) => {
-                            if (focusedWindow) {
-                                focusedWindow.webContents.reload();
-                            } else {
-                                const bw = BrowserWindow.getFocusedWindow();
-                                if (bw) {
-                                    bw.webContents.reload();
-                                } else if (mainWindow) {
-                                    mainWindow.webContents.reload();
-                                } else {
-                                    const arr = BrowserWindow.getAllWindows();
-                                    arr.forEach((bww) => {
-                                        bww.webContents.reload();
-                                    });
-                                }
-                            }
-                        },
-                    },
-                    {
-                        label: "TOGGLE DEV TOOLS",
-                        accelerator: "Alt+CmdOrCtrl+I",
-                        click: (_item: any, focusedWindow: any) => {
-                            if (focusedWindow) {
-                                focusedWindow.webContents.toggleDevTools();
-                            } else {
-                                const bw = BrowserWindow.getFocusedWindow();
-                                if (bw) {
-                                    bw.webContents.toggleDevTools();
-                                } else if (mainWindow) {
-                                    mainWindow.webContents.toggleDevTools();
-                                } else {
-                                    const arr = BrowserWindow.getAllWindows();
-                                    arr.forEach((bww) => {
-                                        bww.webContents.toggleDevTools();
-                                    });
-                                }
-                            }
-                        },
-                    },
-                    {
-                        type: "separator",
-                    },
-                    {
-                        label: "OPEN ALL DEV TOOLS",
-                        accelerator: "Shift+Alt+CmdOrCtrl+I",
-                        click: (_item: any, _focusedWindow: any) => {
-                            const arr = BrowserWindow.getAllWindows();
-                            arr.forEach((bww) => {
-                                bww.webContents.openDevTools({ mode: "detach" });
-                            });
-                        },
-                    },
-                    {
-                        label: "OPEN ALL R2-NAVIGATOR DEV TOOLS",
-                        accelerator: "Shift+Alt+CmdOrCtrl+I",
-                        click: (_item: any, _focusedWindow: any) => {
-                            for (const wc of webContents.getAllWebContents()) {
-                                if (wc.hostWebContents) {
-                                    // wc.hostWebContents.id === readerWindow.webContents.id
-                                    wc.openDevTools({ mode: "detach" });
-                                }
-                            }
-                        },
-                    },
-                ],
-            },
-        );
-    }
-    Menu.setApplicationMenu(Menu.buildFromTemplate(template));
-}
